+++ app/src/main/java/app/revanced/integrations/twitter/Pref.java
@@ -20,6 +20,17 @@ public class Pref {
         return Utils.getStringPref(Settings.VID_SUBFOLDER) + "/" + filename;
     }
 
+    public static int vidMediaHandle() {
+        String val = Utils.getStringPref(Settings.VID_MEDIA_HANDLE);
+        if(val.equals("download_media")){
+            return 1;
+        }
+        if (val.equals("copy_media_link")){
+            return 2;
+        }
+        return 3;
+    }
+
     public static String getSharingLink(String link) {
         String domain = Utils.getStringPref(Settings.CUSTOM_SHARING_DOMAIN);
         return link.replaceFirst("x|twitter", domain);
@@ -111,6 +122,13 @@ public class Pref {
             return 2;
         }
         return 0;
+    }
+
+    public static boolean enableForceTranslate() {
+        return Utils.getBooleanPerf(Settings.TIMELINE_HIDE_FORCE_TRANSLATE);
+    }
+    public static boolean hidePromoteBtn() {
+        return Utils.getBooleanPerf(Settings.TIMELINE_HIDE_PROMOTE_BUTTON);
     }
     public static boolean hideFAB() {
         return Utils.getBooleanPerf(Settings.MISC_HIDE_FAB);