+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsStatus.java
@@ -16,6 +16,8 @@ public class SettingsStatus {
     public static boolean showPollResultsEnabled = false;
     public static boolean browserChooserEnabled = false;
     public static boolean featureFlagsEnabled = false;
+    public static boolean forceTranslate = false;
+    public static boolean hidePromoteButton = false;
 
     public static boolean hideAds = false;
     public static boolean hideGAds = false;
@@ -34,9 +36,16 @@ public class SettingsStatus {
 
     public static boolean profileTabCustomisation = false;
     public static boolean timelineTabCustomisation = false;
+    public static boolean mediaLinkHandle = false;
 
-
+    public static boolean selectableText = false;
+    public static boolean showSensitiveMedia = false;
+    public static boolean enableVidDownload = false;
+    public static void enableVidDownload() { enableVidDownload = true; }
+    public static void showSensitiveMedia() { showSensitiveMedia = true; }
+    public static void selectableText() { selectableText = true; }
     public static void enableDownloadFolder() { changeDownloadEnabled = true; }
+    public static void mediaLinkHandle() { mediaLinkHandle = true; }
     public static void enableCustomSharingDomain() { customSharingDomainEnabled = true; }
     public static void enableFont() { enableFontMod = true; }
     public static void enableFeatureFlags() { featureFlagsEnabled = true; }
@@ -51,6 +60,9 @@ public class SettingsStatus {
     public static void hideLiveThreads() { hideLiveThreads = true; }
     public static void hideBanner() { hideBanner = true; }
     public static void enableShowPollResults() { showPollResultsEnabled = true; }
+    public static void forceTranslate() { forceTranslate = true; }
+    public static void hidePromoteButton() { hidePromoteButton = true; }
+
 
     public static void hideAds() { hideAds = true; }
     public static void hideGAds() { hideGAds = true; }
@@ -70,9 +82,10 @@ public class SettingsStatus {
     public static void profileTabCustomisation() { profileTabCustomisation = true; }
     public static void timelineTabCustomisation() { timelineTabCustomisation = true; }
 
-    public static boolean enableTimelineSection(){ return (hideLiveThreads || hideBanner || hideInlineBmk || showPollResultsEnabled || hideImmersivePlayer); }
-    public static boolean enableMiscSection() { return (enableFontMod || hideRecommendedUsers || hideFAB || hideViewCount || hideCommunityNote || customSharingDomainEnabled || hideFABBtns); }
+    public static boolean enableTimelineSection(){ return (forceTranslate || hidePromoteButton || hideCommunityNote|| hideLiveThreads || hideBanner || hideInlineBmk || showPollResultsEnabled || hideImmersivePlayer); }
+    public static boolean enableMiscSection() { return (enableFontMod || hideRecommendedUsers || hideFAB || hideViewCount || customSharingDomainEnabled || hideFABBtns); }
     public static boolean enableAdsSection() {return (hideAds|| hideGAds || hideWTF || hideCTS || hideCTJ || hideDetailedPosts || hideRBMK ||hidePromotedTrend); }
+    public static boolean enableDownloadSection() {return (changeDownloadEnabled || mediaLinkHandle); }
 
     public static boolean enablePremiumSection() {return (enableReaderMode || enableUndoPosts || enableAppIconNNavIcon); }
     public static boolean enableCustomisationSection() {return (profileTabCustomisation || timelineTabCustomisation); }